# Optimize: Production-mode `react` in manager